Release version v0.6.3
Metacello new
    baseline: 'ThreadedFFI';
    repository: 'github://feenkcom/threadedFFI-Plugin:v0.6.3';
    load
All commits (inlcuding upstream repositories) since last build:
https://github.com/feenkcom/threadedFFI-Plugin/commit/596fe2 by Alistair Grant
MCGitBasedNetworkRepository>>createIcebergRepositoryWithFallbackFor:url: bug fix

If plaintext credentials are already available, retry the first time an exception is raised.

https://github.com/feenkcom/threadedFFI-Plugin/commit/214ddb by Alistair Grant
Metacello: Prompt for user credentials if https is being used

Metacello currently retries to load packages using https if ssh credentials fail.  It relies on Iceberg prompting the user for the credentials.  However libgit does this in a callback.  With Alien callbacks it works as the callback is run from the UI thread, however in threaded FFI the callback is run in a separate thread, leaving the UI thread blocked and unable to handle user input.

- Modify Metacello to handle invalid credentials, prompt the user if they are missing, and then retry.
- When enabling threaded FFI disable Iceberg's interactive credentials provider so that the system doesn't lock up.